chore: Move CVE revision endpoint to endpoints section of app
